+++ src/commonMain/kotlin/csense/kotlin/extensions/Nullability.kt
@@ -5,6 +5,7 @@ package csense.kotlin.extensions
 import csense.kotlin.EmptyFunction
 import csense.kotlin.FunctionUnit
 import csense.kotlin.extensions.primitives.ifTrue
+import kotlin.contracts.*
 
 
 /**
@@ -12,7 +13,11 @@ import csense.kotlin.extensions.primitives.ifTrue
  * @receiver Any? the optional value
  * @param action EmptyFunction the function to run if the receiver is not null
  */
+@OptIn(ExperimentalContracts::class)
 inline fun <T> T?.ifNull(action: EmptyFunction) {
+    contract {
+        callsInPlace(action,InvocationKind.AT_MOST_ONCE)
+    }
     this.isNull.ifTrue(action)
 }
 
@@ -21,7 +26,11 @@ inline fun <T> T?.ifNull(action: EmptyFunction) {
  * @receiver T? the optional value
  * @param action FunctionUnit<T> the action to call if the receiver is not null
  */
+@OptIn(ExperimentalContracts::class)
 inline fun <T> T?.ifNotNull(action: FunctionUnit<T>) {
+    contract {
+        callsInPlace(action,InvocationKind.AT_MOST_ONCE)
+    }
     this?.let(action)
 }