+++ src/commonTest/kotlin/csense/kotlin/extensions/collections/IterableTest.kt
@@ -4,8 +4,7 @@ import csense.kotlin.tests.assertions.*
 import kotlin.test.*
 
 class IterableTest {
-
-
+    
     @Test
     fun iterableTIsEmpty() {
         val empty: Iterable<String> = arrayListOf()
@@ -15,7 +14,7 @@ class IterableTest {
         val multiple: Iterable<String> = arrayListOf("a", "b")
         multiple.isEmpty().assertFalse()
     }
-
+    
     @Test
     fun iterableTIsNotEmpty() {
         val empty: Iterable<String> = arrayListOf()
@@ -24,52 +23,52 @@ class IterableTest {
         single.isNotEmpty().assertTrue()
         val multiple: Iterable<String> = arrayListOf("a", "b")
         multiple.isNotEmpty().assertTrue()
-
+        
     }
-
+    
     @Test
     fun iterableTSkipIfEmptyOr() {
         val empty: Iterable<String> = arrayListOf()
         empty.skipIfEmptyOr { shouldNotBeCalled() }
-
+        
         val single: Iterable<String> = arrayListOf("1")
         assertCalled { single.skipIfEmptyOr(it) }
-
+        
         val multiple: Iterable<String> = arrayListOf("a", "b")
         assertCalled { multiple.skipIfEmptyOr(it) }
     }
-
-    class iterableEForEachNotNull {
+    
+    class IterableEForEachNotNull {
         @Test
         fun empty() {
             val itt: Iterable<String?> = listOf()
             itt.forEachNotNull { shouldNotBeCalled() }
         }
-
+        
         @Test
         fun singelNull() {
             val itt: Iterable<String?> = listOf(null)
             itt.forEachNotNull { shouldNotBeCalled() }
         }
-
+        
         @Test
         fun singleNotNull() = assertCalled {
             val itt: Iterable<String?> = listOf("asd")
             itt.forEachNotNull { it() }
         }
-
+        
         @Test
         fun multiple() = assertCalled(times = 2) {
             val itt: Iterable<String?> = listOf("asd", "1234")
             itt.forEachNotNull { it() }
         }
-
+        
         @Test
         fun multipleNull() = assertNotCalled {
             val itt: Iterable<String?> = listOf(null, null)
             itt.forEachNotNull { it() }
         }
-
+        
         @Test
         fun multipleMixed() = assertCalled(times = 3) {
             val itt: Iterable<String?> =
@@ -77,43 +76,43 @@ class IterableTest {
             itt.forEachNotNull { it() }
         }
     }
-
+    
     //region Invoke each lazy (1 - 6 )
     class IterableFunction1I1OInvokeEachWithLazyElement {
         @Test
         fun empty() {
             listOf<Function1<String, String>>().invokeEachWithLazy(::shouldNotBeCalled)
         }
-
+        
         @Test
         fun single() {
             //TODO test single element condition here.
         }
-
+        
         @Test
         fun multiple() {
             //TODO test multiple element condition here.
         }
     }
-
+    
     class IterableFunction2I1I2OInvokeEachWithLazyFirstElement {
         @Test
         fun empty() {
             listOf<Function2<String, String, String>>().invokeEachWithLazy(::shouldNotBeCalled, ::shouldNotBeCalled)
             //TODO test empty condition here.
         }
-
+        
         @Test
         fun single() {
             //TODO test single element condition here.
         }
-
+        
         @Test
         fun multiple() {
             //TODO test multiple element condition here.
         }
     }
-
+    
     class IterableFunction3I1I2I3OInvokeEachWithLazyFirstElement {
         @Test
         fun empty() {
@@ -121,54 +120,54 @@ class IterableTest {
                     ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled)
             //TODO test empty condition here.
         }
-
+        
         @Test
         fun single() {
             //TODO test single element condition here.
         }
-
+        
         @Test
         fun multiple() {
             //TODO test multiple element condition here.
         }
     }
-
+    
     class IterableFunction4I1I2I3I4OInvokeEachWithLazyFirstElement {
         @Test
         fun empty() {
             listOf<Function4<String, String, String, String, String>>().invokeEachWithLazy(
                     ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled)
         }
-
+        
         @Test
         fun single() {
             //TODO test single element condition here.
         }
-
+        
         @Test
         fun multiple() {
             //TODO test multiple element condition here.
         }
     }
-
+    
     class IterableFunction5I1I2I3I4I5OInvokeEachWithLazyFirstElement {
         @Test
         fun empty() {
             listOf<Function5<String, String, String, String, String, String>>().invokeEachWithLazy(
                     ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled)
         }
-
+        
         @Test
         fun single() {
             //TODO test single element condition here.
         }
-
+        
         @Test
         fun multiple() {
             //TODO test multiple element condition here.
         }
     }
-
+    
     class IterableFunction6I1I2I3I4I5I6OInvokeEachWithLazyFirstElement {
         @Test
         fun empty() {
@@ -176,19 +175,18 @@ class IterableTest {
                     ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled,
                     ::shouldNotBeCalled, ::shouldNotBeCalled, ::shouldNotBeCalled)
         }
-
+        
         @Test
         fun single() {
             //TODO test single element condition here.
         }
-
+        
         @Test
         fun multiple() {
             //TODO test multiple element condition here.
         }
     }
     //endregion
-    
     
     
     class IterableForEachIsInstanceAction {
@@ -230,6 +228,77 @@ class IterableTest {
                 shouldNotBeCalled()
             }
             anyList.forEachIsInstance<Array<*>> { shouldNotBeCalled() }
+        }
+    }
+    
+    class IterableEPartitionSafe {
+        @Test
+        fun empty() {
+            listOf<String>().partitionSafe { shouldNotBeCalled() }
+        }
+        
+        @Test
+        fun single() {
+            listOf(1).partitionSafe { true }.apply {
+                trueForPredicate.assertSize(1)
+                falseForPredicate.assertEmpty()
+                trueForPredicate.first().assert(1)
+            }
+            listOf(2).partitionSafe { false }.apply {
+                trueForPredicate.assertEmpty()
+                falseForPredicate.assertSize(1)
+                falseForPredicate.first().assert(2)
+            }
+            listOf(3).partitionSafe { it == 3 }.apply {
+                trueForPredicate.assertSize(1)
+                falseForPredicate.assertEmpty()
+            }
+        }
+        
+        @Test
+        fun multiple() {
+            listOf("test", "1234", "a").partitionSafe { false }.apply {
+                trueForPredicate.assertEmpty()
+                falseForPredicate.assertSize(3)
+            }
+            listOf("test", "1234", "a").partitionSafe { true }.apply {
+                trueForPredicate.assertSize(3)
+                falseForPredicate.assertEmpty()
+            }
+            listOf("test", "1234", "a").partitionSafe { it == "a" }.apply {
+                trueForPredicate.assertSize(1)
+                falseForPredicate.assertSize(2)
+                trueForPredicate.first().assert("a")
+                falseForPredicate.assertContainsAll("test", "1234")
+            }
+            listOf("test", "1234", "a").partitionSafe { it.length == 4 }.apply {
+                trueForPredicate.assertSize(2)
+                falseForPredicate.assertSize(1)
+                trueForPredicate.assertContainsAll("1234", "test")
+                falseForPredicate.first().assert("a")
+            }
+        }
+    }
+    
+    class IterableELargest {
+        @Test
+        fun empty() {
+            listOf<String>().largest { shouldNotBeCalled() }
+        }
+        
+        @Test
+        fun single() {
+            listOf("test")
+                    .largest {
+                        it.assert("test")
+                        it.length
+                    }.assertNotNullAndEquals("test")
+        }
+        
+        @Test
+        fun multiple() {
+            listOf(42, 100, 1, 102).largest { it }.assertNotNullAndEquals(102)
+            listOf(42, 100, 1, 102).largest { 0 - it }.assertNotNullAndEquals(1)
         }
     }
 }